feet: commit

+++ database/migrations/2024_07_30_133608_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('external_code')->unique();
             $table->string('name');
             $table->text('description');
-            $table->unsignedBigInteger('price');
+            $table->decimal('price');
             $table->string('discount');
             $table->timestamps();
         });
